Add new components and data structures for contact, loading, footer, background elements, social links, education, responsibilities, skills, and projects sections
