+++ data/education.ts
@@ -1,0 +1,26 @@
+export const education = [
+  {
+    institution: "Lalbhai Dalpatbhai College of Engineering",
+    degree: "Bachelor of Engineering in Electronics & Communications",
+    period: "2021-2025",
+    website: "https://ldce.ac.in",
+    skills: [
+      "VLSI Design",
+      "Digital Electronics",
+      "Embedded Systems",
+      "Circuit Theory",
+      "Internet of Things",
+      "Computer Networks",
+      "Microcontrollers",
+      "Machine Learning",
+      "Signal Processing",
+    ],
+  },
+  {
+    institution: "The H. B. Kapadia New High School",
+    degree: "Higher Secondary Education",
+    period: "2021-2022",
+    website: "https://hbkapadia.com/",
+    skills: ["Physics", "Mathematics", "Chemistry", "Computer Science"],
+  },
+];
