+++ data/projects.ts
@@ -1,0 +1,51 @@
+export const projects = [
+  {
+    title: "Gesture-Controlled Car",
+    image: "/gesture.gif",
+    description:
+      "Developed a gesture-controlled system interpreting real-time hand gestures to wirelessly command the car's movements via ESP-NOW protocol, ensuring low-latency communication. This project represents the second phase of the Wi-Fi car series, enhancing control capabilities through gesture recognition.",
+    techStack: ["IoT", "Wireless Communication", "Embedded Systems"],
+    githubLink: "https://github.com/ayushilathiya/Gesture_Controlled_Car",
+    demoLink: "https://youtu.be/n70FIHps4CA",
+  },
+  {
+    title: "Live ECG Monitoring System",
+    image: "/ecgmoni.gif",
+    description:
+      "ESP-based IoT ECG monitoring system utilizing the ECG sensor for real-time cardiac signal acquisition and wireless transmission. Implements signal filtering for noise reduction and cloud integration for remote monitoring via ThingSpeak. Ensures low-latency data transfer and precise ECG waveform analysis.",
+    techStack: [
+      "Real-time Data Acquisition",
+      "Thingspeak API",
+      "MIT App Inventor",
+    ],
+    githubLink: "https://github.com/ayushilathiya/ECG-Monitoring",
+    demoLink: "https://youtu.be/EAjrd2bCG9A",
+  },
+  {
+    title: "WiFi-Controlled Car",
+    image: "/wifi.gif",
+    description:
+      "Engineered a wireless-controlled vehicle using ESP8266 NodeMCU, featuring real-time control through a custom mobile interface. Implemented smooth motion control and responsive directional changes with L298N motor driver integration.",
+    techStack: ["ESP8266", "JavaScript", "HTML", "CSS", "IoT"],
+    githubLink: "https://github.com/ayushilathiya/WiFi-Controlled-Car",
+    demoLink: "https://youtu.be/vaGgphrMhso",
+  },
+  {
+    title: "3D Modeling Using Sensor-Driven 3D Visualization",
+    image: "/mpu6050.gif",
+    description:
+      "Created an innovative system that transforms real-time sensor data into dynamic 3D visualizations. Integrated MPU6050 sensor data with WebGL rendering to achieve precise spatial mapping and interactive model manipulation.",
+    techStack: ["MPU6050", "WebGL", "Three.js", "Data Visualization"],
+    githubLink: "https://github.com/ayushilathiya/MPU6050-3D-Visualization",
+    demoLink: "https://ayushilathiya.hashnode.dev/3d-modeling-sensors",
+  },
+  {
+    title: "ESP-NOW Protocol",
+    image: "/espnow.gif",
+    description:
+      "Implemented a robust wireless communication system using ESP-NOW protocol, achieving low-latency data transfer between multiple ESP8266 modules. Optimized for reliable peer-to-peer communication with minimal power consumption.",
+    techStack: ["C++", "ESP-NOW", "Wireless Communication"],
+    githubLink: "https://github.com/ayushilathiya/ESP-NOW-Protocol",
+    demoLink: "https://youtu.be/NGjMKT3Scls",
+  },
+];
