+++ data/responsibilities.ts
@@ -1,0 +1,15 @@
+export const responsibilities = [
+  {
+    title: "Google Developer Student Club - LDCE",
+    org: "Team Member",
+    period: "Nov 2023 - Oct 2024",
+    desc: "Curated technical content for various GDSC events, ensuring engaging and informative sessions for attendees. Gained hands-on experience with Android Studio, Open Source Contributions, and Machine Learning Basics through TensorFlow events and the Google Cloud campaign. Actively participated in organizing and hosting tech workshops, fostering a learning environment for peers.",
+    skills: [
+      "Technical Writing",
+      "Android Studio",
+      "Open Source",
+      "Google Cloud",
+    ],
+    link: "https://in.linkedin.com/company/gdsc-ldce",
+  },
+];
