+++ data/skills.ts
@@ -1,0 +1,61 @@
+export const skills = [
+  {
+    name: "MOSFETs",
+    url: "https://learn.sparkfun.com/tutorials/transistors/all",
+  },
+  { name: "Python", url: "https://www.python.org/doc/" },
+  {
+    name: "Verilog",
+    url: "https://www.chipverify.com/verilog/verilog-tutorial",
+  },
+  { name: "IoT", url: "https://www.arduino.cc/education/courses/learning-iot" },
+  {
+    name: "Xilinx",
+    url: "https://www.xilinx.com/products/design-tools/vivado.html",
+  },
+  {
+    name: "Machine Learning",
+    url: "https://developers.google.com/machine-learning",
+  },
+  { name: "NumPy", url: "https://numpy.org/doc/stable/" },
+  { name: "MicroPython", url: "https://micropython.org/documentation" },
+  { name: "C", url: "https://devdocs.io/c/" },
+  { name: "C++", url: "https://cplusplus.com/doc/tutorial/" },
+  {
+    name: "Embedded C",
+    url: "https://www.embedded.com/collection/programming-languages/",
+  },
+  { name: "Vercel", url: "https://vercel.com/docs" },
+  { name: "TensorFlow", url: "https://www.tensorflow.org/learn" },
+  { name: "Proteus", url: "https://www.labcenter.com/documentation/" },
+  { name: "HTML", url: "https://developer.mozilla.org/en-US/docs/Web/HTML" },
+  { name: "CSS", url: "https://developer.mozilla.org/en-US/docs/Web/CSS" },
+  {
+    name: "JavaScript",
+    url: "https://developer.mozilla.org/en-US/docs/Web/JavaScript",
+  },
+  { name: "TypeScript", url: "https://www.typescriptlang.org/docs/" },
+  { name: "Tailwind CSS", url: "https://tailwindcss.com/docs" },
+  {
+    name: "Matplotlib",
+    url: "https://matplotlib.org/stable/tutorials/index.html",
+  },
+  { name: "Raspberry Pi", url: "https://www.raspberrypi.com/documentation/" },
+  { name: "Firebase", url: "https://firebase.google.com/docs" },
+  {
+    name: "FPGA",
+    url: "https://www.intel.com/content/www/us/en/products/details/fpga/development-tools.html",
+  },
+  {
+    name: "Signal Processing",
+    url: "https://www.analog.com/en/design-center/landing-pages/001/beginners-guide-to-dsp.html",
+  },
+  {
+    name: "LTSpice",
+    url: "https://www.analog.com/en/design-center/design-tools-and-calculators/ltspice-simulator.html",
+  },
+  {
+    name: "Computer Networks",
+    url: "https://www.cisco.com/c/en/us/solutions/enterprise-networks/what-is-computer-networking.html",
+  },
+];
